requirements.txt
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
traceloop-sdk==0.17.1
arize-phoenix==1.2.1
streamlit==1.31.1
requests==2.28.2
Pillow==9.5.0
langchain
langchain_openai
python-dotenv
pandas
numpy
